https://issues.apache.org/jira/browse/SOLR-15286
Added log to the HealthCheckHandler for legacy mode to report followers healthy based on the index generation diff
